requirements.txt
streamlit
openai
python-dotenv
pypdf
opencv-python
pillow
